FIX: Loading kanban lists when returning from topic view
In this situation, core's discovery.hbs initially renders the outlet inside a `display: none` element which means that the IntersectionObserver does not fire. Moving the intersection observer setup to the next runloop iteration guarantees that the `display: none` will have been removed, and therefore the IntersectionObserver will fire and load topics as expected.